update pytorch example
