+++ integrations-and-supported-tools/pytorch/scripts/requirements.txt
@@ -2,3 +2,4 @@ neptune
 numpy
 torch
 torchvision
+torchviz
