+++ integrations-and-supported-tools/pytorch/scripts/run_examples.sh
@@ -3,8 +3,5 @@ set -e
 echo "Installing requirements..."
 pip install -U -r requirements.txt
 
-echo "Running Neptune_Pytorch_Support.py..."
-python Neptune_Pytorch_Support.py
-
-echo "Running Neptune_Pytorch_Support_more_options.py..."
-python Neptune_Pytorch_Support_more_options.py
+echo "Running Neptune_Pytorch.py..."
+python Neptune_Pytorch.py
